@@ -76,7 +76,7 @@
 			<NcEmptyContent
 				v-if="loading"
 				class="forms-emptycontent"
-				:name="t('forms', 'Loading forms …')">
+				:name="t('forms', 'Loading forms …')">
 				<template #icon>
 					<NcLoadingIcon :size="64" />
 				</template>
@@ -118,25 +118,27 @@
 		<!-- No errors show router content -->
 		<template v-else>
 			<router-view
-				:form.sync="selectedForm"
-				:sidebar-opened.sync="sidebarOpened"
+				v-model:form="selectedForm"
+				v-model:sidebar-opened="sidebarOpened"
 				@open-sharing="openSharing" />
 			<Sidebar
 				v-if="!selectedForm.partial && canEdit"
 				:form="selectedForm"
-				:sidebar-opened.sync="sidebarOpened"
-				:active.sync="sidebarActive" />
+				v-model:sidebar-opened="sidebarOpened"
+				v-model:active="sidebarActive" />
 		</template>
 
 		<!-- Archived forms modal -->
 		<ArchivedFormsModal
-			:open.sync="showArchivedForms"
+			v-model:open="showArchivedForms"
 			:forms="archivedForms"
 			@clone="onCloneForm" />
 	</NcContent>
 </template>
 
 <script>
+import { ref, computed, onMounted, onUnmounted } from 'vue'
+import { useRoute, useRouter } from 'vue-router'
 import axios from '@nextcloud/axios'
 import { showError } from '@nextcloud/dialogs'
 import { emit, subscribe, unsubscribe } from '@nextcloud/event-bus'
@@ -163,8 +165,9 @@ import { FormState } from './models/Constants.ts'
 import logger from './utils/Logger.js'
 import OcsResponse2Data from './utils/OcsResponse2Data.js'
 
+const appName = 'forms'
+
 export default {
-	// eslint-disable-next-line vue/multi-word-component-names
 	name: 'Forms',
 
 	components: {
@@ -184,179 +187,121 @@ export default {
 		Sidebar,
 	},
 
-	mixins: [PermissionTypes],
-
 	setup() {
-		return {
-			isMobile: useIsMobile(),
-		}
-	},
+		const route = useRoute()
+		const router = useRouter()
+		const isMobile = useIsMobile()
 
-	data() {
-		return {
-			loading: true,
-			sidebarOpened: false,
-			sidebarActive: 'forms-sharing',
-			forms: [],
-			allSharedForms: [],
+		const loading = ref(true)
+		const sidebarOpened = ref(false)
+		const sidebarActive = ref('forms-sharing')
+		const forms = ref([])
+		const allSharedForms = ref([])
+		const showArchivedForms = ref(false)
+		const canCreateForms = ref(loadState(appName, 'appConfig').canCreateForms)
 
-			showArchivedForms: false,
+		const PERMISSION_TYPES = PermissionTypes.data().PERMISSION_TYPES
 
-			canCreateForms: loadState(appName, 'appConfig').canCreateForms,
-		}
-	},
-
-	computed: {
-		canEdit() {
-			return this.selectedForm.permissions.includes(
-				this.PERMISSION_TYPES.PERMISSION_EDIT,
-			)
-		},
-
-		hasForms() {
-			return this.allSharedForms.length > 0 || this.forms.length > 0
-		},
-
-		/**
-		 * All own active forms
-		 */
-		ownedForms() {
-			return this.forms.filter((form) => form.state !== FormState.FormArchived)
-		},
-
-		/**
-		 * All active shared forms
-		 */
-		sharedForms() {
-			return this.allSharedForms.filter(
-				(form) => form.state !== FormState.FormArchived,
-			)
-		},
-
-		/**
-		 * All forms that have been archived
-		 */
-		archivedForms() {
-			return [...this.forms, ...this.allSharedForms].filter(
-				(form) => form.state === FormState.FormArchived,
-			)
-		},
-
-		routeHash() {
-			return this.$route.params.hash
-		},
-
-		// If the user is allowed to access this route
-		routeAllowed() {
-			// Check formId from initial state on app initialization
-			if (this.loading && loadState(appName, 'formId') === 'invalid') {
-				return false
-			}
-
-			// Not allowed, if no hash
-			if (!this.routeHash) {
-				return false
-			}
-
-			// Try to find form in owned & shared list
-			const form = [...this.forms, ...this.allSharedForms].find(
-				(form) => form.hash === this.routeHash,
-			)
-
-			// If no form found, load it from server. Route will be automatically re-evaluated.
-			if (form === undefined) {
-				this.fetchPartialForm(this.routeHash)
-				return false
-			}
-
-			// Return whether route is in the permissions-list
-			return form?.permissions.includes(this.$route.name)
-		},
-
-		selectedForm: {
-			get() {
-				if (this.routeAllowed) {
-					return this.forms
-						.concat(this.allSharedForms)
-						.find((form) => form.hash === this.routeHash)
+		const selectedForm = computed({
+			get: () => {
+				if (routeAllowed.value) {
+					return [...forms.value, ...allSharedForms.value].find(
+						(form) => form.hash === routeHash.value,
+					) || {}
 				}
 				return {}
 			},
+			set: (form) => {
+				sidebarOpened.value = false
 
-			set(form) {
-				// always close sidebar
-				this.sidebarOpened = false
-
-				// If a owned form
-				let index = this.forms.findIndex(
-					(search) => search.hash === this.routeHash,
+				let index = forms.value.findIndex(
+					(search) => search.hash === routeHash.value,
 				)
 				if (index > -1) {
-					this.$set(this.forms, index, form)
+					forms.value[index] = form
 					return
 				}
-				// Otherwise a shared form
-				index = this.allSharedForms.findIndex(
-					(search) => search.hash === this.routeHash,
+
+				index = allSharedForms.value.findIndex(
+					(search) => search.hash === routeHash.value,
 				)
 				if (index > -1) {
-					this.$set(this.allSharedForms, index, form)
+					allSharedForms.value[index] = form
 				}
 			},
-		},
-	},
+		})
 
-	beforeMount() {
-		this.loadForms()
-	},
+		const canEdit = computed(() => {
+			return selectedForm.value.permissions?.includes(
+				PERMISSION_TYPES.PERMISSION_EDIT,
+			)
+		})
 
-	mounted() {
-		subscribe('forms:last-updated:set', (id) => this.onLastUpdatedByEventBus(id))
-		subscribe('forms:ownership-transfered', (id) => this.onDeleteForm(id))
-	},
+		const hasForms = computed(() => {
+			return allSharedForms.value.length > 0 || forms.value.length > 0
+		})
 
-	unmounted() {
-		unsubscribe('forms:last-updated:set', (id) =>
-			this.onLastUpdatedByEventBus(id),
-		)
-		unsubscribe('forms:ownership-transfered', (id) => this.onDeleteForm(id))
-	},
+		const ownedForms = computed(() => {
+			return forms.value.filter((form) => form.state !== FormState.FormArchived)
+		})
 
-	methods: {
-		/**
-		 * Closes the App-Navigation on mobile-devices
-		 */
-		mobileCloseNavigation() {
-			if (this.isMobile) {
+		const sharedForms = computed(() => {
+			return allSharedForms.value.filter(
+				(form) => form.state !== FormState.FormArchived,
+			)
+		})
+
+		const archivedForms = computed(() => {
+			return [...forms.value, ...allSharedForms.value].filter(
+				(form) => form.state === FormState.FormArchived,
+			)
+		})
+
+		const routeHash = computed(() => route.params.hash)
+
+		const routeAllowed = computed(() => {
+			if (loading.value && loadState(appName, 'formId') === 'invalid') {
+				return false
+			}
+
+			if (!routeHash.value) {
+				return false
+			}
+
+			const form = [...forms.value, ...allSharedForms.value].find(
+				(form) => form.hash === routeHash.value,
+			)
+
+			if (form === undefined) {
+				fetchPartialForm(routeHash.value)
+				return false
+			}
+
+			return form?.permissions.includes(route.name)
+		})
+
+		const mobileCloseNavigation = () => {
+			if (isMobile.value) {
 				emit('toggle-navigation', { open: false })
 			}
-		},
+		}
 
-		/**
-		 * Open a form and its sidebar for sharing
-		 *
-		 * @param {string} hash the hash of the form to load
-		 */
-		openSharing(hash) {
-			if (hash !== this.routeHash) {
-				this.$router.push({ name: 'edit', params: { hash } })
+		const openSharing = (hash) => {
+			if (hash !== routeHash.value) {
+				router.push({ name: 'edit', params: { hash } })
 			}
-			this.sidebarActive = 'forms-sharing'
-			this.sidebarOpened = true
-		},
+			sidebarActive.value = 'forms-sharing'
+			sidebarOpened.value = true
+		}
 
-		/**
-		 * Initial forms load
-		 */
-		async loadForms() {
-			this.loading = true
+		const loadForms = async () => {
+			loading.value = true
 
-			// Load Owned forms
 			try {
 				const response = await axios.get(
 					generateOcsUrl('apps/forms/api/v3/forms'),
 				)
-				this.forms = OcsResponse2Data(response)
+				forms.value = OcsResponse2Data(response)
 			} catch (error) {
 				logger.error('Error while loading owned forms list', { error })
 				showError(
@@ -364,12 +309,11 @@ export default {
 				)
 			}
 
-			// Load shared forms
 			try {
 				const response = await axios.get(
 					generateOcsUrl('apps/forms/api/v3/forms?type=shared'),
 				)
-				this.allSharedForms = OcsResponse2Data(response)
+				allSharedForms.value = OcsResponse2Data(response)
 			} catch (error) {
 				logger.error('Error while loading shared forms list', {
 					error,
@@ -379,18 +323,13 @@ export default {
 				)
 			}
 
-			this.loading = false
-		},
+			loading.value = false
+		}
 
-		/**
-		 * Fetch a partial form by its hash and add it to the shared forms list.
-		 *
-		 * @param {string} hash the hash of the form to load
-		 */
-		async fetchPartialForm(hash) {
+		const fetchPartialForm = async (hash) => {
 			await new Promise((resolve) => {
 				const wait = () => {
-					if (this.loading) {
+					if (loading.value) {
 						window.setTimeout(wait, 250)
 					} else {
 						resolve()
@@ -399,9 +338,9 @@ export default {
 				wait()
 			})
 
-			this.loading = true
+			loading.value = true
 			if (
-				[...this.forms, ...this.allSharedForms].find(
+				[...forms.value, ...allSharedForms.value].find(
 					(form) => form.hash === hash,
 				) === undefined
 			) {
@@ -413,54 +352,47 @@ export default {
 					)
 					const form = OcsResponse2Data(response)
 
-					// If the user has (at least) submission-permissions, add it to the shared forms
 					if (
 						form.permissions.includes(
-							this.PERMISSION_TYPES.PERMISSION_SUBMIT,
+							PERMISSION_TYPES.PERMISSION_SUBMIT,
 						)
 					) {
-						this.allSharedForms.push(form)
+						allSharedForms.value.push(form)
 					}
 				} catch (error) {
 					logger.error(`Form ${hash} not found`, { error })
 					showError(t('forms', 'Form not found'))
 
 					if ([403, 404].includes(error.response?.status)) {
-						if (this.$route.name !== 'root') {
-							this.$router.push({ name: 'root' })
+						if (route.name !== 'root') {
+							router.push({ name: 'root' })
 						}
 					}
 				}
 			}
 
-			this.loading = false
-		},
+			loading.value = false
+		}
 
-		async onNewForm() {
+		const onNewForm = async () => {
 			try {
-				// Request a new empty form
 				const response = await axios.post(
 					generateOcsUrl('apps/forms/api/v3/forms'),
 				)
 				const newForm = OcsResponse2Data(response)
-				this.forms.unshift(newForm)
-				this.$router.push({
+				forms.value.unshift(newForm)
+				router.push({
 					name: 'edit',
 					params: { hash: newForm.hash },
 				})
-				this.mobileCloseNavigation()
+				mobileCloseNavigation()
 			} catch (error) {
 				logger.error('Unable to create new form', { error })
 				showError(t('forms', 'Unable to create a new form'))
 			}
-		},
+		}
 
-		/**
-		 * Request to clone a form, store returned form and open it.
-		 *
-		 * @param {number} id id of the form to clone
-		 */
-		async onCloneForm(id) {
+		const onCloneForm = async (id) => {
 			try {
 				const response = await axios.post(
 					generateOcsUrl('apps/forms/api/v3/forms?fromId={id}', {
@@ -468,53 +400,80 @@ export default {
 					}),
 				)
 				const newForm = OcsResponse2Data(response)
-				this.forms.unshift(newForm)
-				this.$router.push({
+				forms.value.unshift(newForm)
+				router.push({
 					name: 'edit',
 					params: { hash: newForm.hash },
 				})
-				this.mobileCloseNavigation()
+				mobileCloseNavigation()
 			} catch (error) {
 				logger.error(`Unable to copy form ${id}`, { error })
 				showError(t('forms', 'Unable to copy form'))
 			}
-		},
+		}
 
-		/**
-		 * Remove form from forms list after successful server deletion request
-		 *
-		 * @param {number} id the form id
-		 */
-		async onDeleteForm(id) {
-			const formIndex = this.forms.findIndex((form) => form.id === id)
-			const deletedHash = this.forms[formIndex].hash
+		const onDeleteForm = async (id) => {
+			const formIndex = forms.value.findIndex((form) => form.id === id)
+			const deletedHash = forms.value[formIndex].hash
 
-			this.forms.splice(formIndex, 1)
+			forms.value.splice(formIndex, 1)
 
-			// Redirect if current form has been deleted
-			if (deletedHash === this.routeHash && this.$route.name !== 'root') {
-				this.$router.push({ name: 'root' })
+			if (deletedHash === routeHash.value && route.name !== 'root') {
+				router.push({ name: 'root' })
 			}
-		},
+		}
 
-		/**
-		 * Update last updated timestamp in given form
-		 *
-		 * @param {number} id the form id
-		 */
-		onLastUpdatedByEventBus(id) {
-			const formIndex = this.forms.findIndex((form) => form.id === id)
+		const onLastUpdatedByEventBus = (id) => {
+			const formIndex = forms.value.findIndex((form) => form.id === id)
 			if (formIndex !== -1) {
-				this.forms[formIndex].lastUpdated = moment().unix()
-				this.forms.sort((b, a) => a.lastUpdated - b.lastUpdated)
+				forms.value[formIndex].lastUpdated = moment().unix()
+				forms.value.sort((b, a) => a.lastUpdated - b.lastUpdated)
 			} else {
-				const sharedFormIndex = this.allSharedForms.findIndex(
+				const sharedFormIndex = allSharedForms.value.findIndex(
 					(form) => form.id === id,
 				)
-				this.allSharedForms[sharedFormIndex].lastUpdated = moment().unix()
-				this.allSharedForms.sort((b, a) => a.lastUpdated - b.lastUpdated)
+				allSharedForms.value[sharedFormIndex].lastUpdated = moment().unix()
+				allSharedForms.value.sort((b, a) => a.lastUpdated - b.lastUpdated)
 			}
-		},
+		}
+
+		onMounted(() => {
+			loadForms()
+			subscribe('forms:last-updated:set', onLastUpdatedByEventBus)
+			subscribe('forms:ownership-transfered', onDeleteForm)
+		})
+
+		onUnmounted(() => {
+			unsubscribe('forms:last-updated:set', onLastUpdatedByEventBus)
+			unsubscribe('forms:ownership-transfered', onDeleteForm)
+		})
+
+		return {
+			loading,
+			sidebarOpened,
+			sidebarActive,
+			forms,
+			allSharedForms,
+			showArchivedForms,
+			canCreateForms,
+			isMobile,
+			selectedForm,
+			canEdit,
+			hasForms,
+			ownedForms,
+			sharedForms,
+			archivedForms,
+			routeHash,
+			routeAllowed,
+			mobileCloseNavigation,
+			openSharing,
+			loadForms,
+			fetchPartialForm,
+			onNewForm,
+			onCloneForm,
+			onDeleteForm,
+			onLastUpdatedByEventBus,
+		}
 	},
 }
 </script>
@@ -526,12 +485,10 @@ export default {
 	padding: var(--app-navigation-padding);
 }
 
-// Fix the margin of the lists
 .forms-navigation__list-heading {
 	margin-block: calc(var(--default-grid-baseline) * 2) 0 !important;
 
 	:deep(h2) {
-		// Make the list more condensed
 		margin-block: 0;
 	}
 }
