@@ -2,17 +2,17 @@
  * SPDX-FileCopyrightText: 2022 Nextcloud GmbH and Nextcloud contributors
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
+import { createApp } from 'vue'
 import { translate, translatePlural } from '@nextcloud/l10n'
-import Vue from 'vue'
 import FormsEmptyContent from './FormsEmptyContent.vue'
 
 import 'vite/modulepreload-polyfill'
 
-Vue.prototype.t = translate
-Vue.prototype.n = translatePlural
+const app = createApp(FormsEmptyContent)
 
-export default new Vue({
-	el: '#content',
-	name: 'FormsEmptyContent',
-	render: (h) => h(FormsEmptyContent),
-})
+app.config.globalProperties.t = translate
+app.config.globalProperties.n = translatePlural
+
+app.mount('#content')
+
+export default app
