@@ -3,23 +3,21 @@
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 
+import { createApp } from 'vue'
 import { translate, translatePlural } from '@nextcloud/l10n'
-import Tooltip from '@nextcloud/vue/directives/Tooltip'
-import Vue from 'vue'
 import Forms from './Forms.vue'
 import router from './router.js'
 
 import 'vite/modulepreload-polyfill'
 import '@nextcloud/dialogs/style.css'
 
-Vue.directive('tooltip', Tooltip)
+const app = createApp(Forms)
 
-Vue.prototype.t = translate
-Vue.prototype.n = translatePlural
+app.use(router)
 
-export default new Vue({
-	el: '#content',
-	name: 'FormsRoot',
-	router,
-	render: (h) => h(Forms),
-})
+app.config.globalProperties.t = translate
+app.config.globalProperties.n = translatePlural
+
+app.mount('#content')
+
+export default app
